requirements.txt
aiofiles==23.2.1
aiohttp==3.10.11
curl_cffi==0.7.3
google-genai>=1.0.0
groq>=0.9.0
python-dotenv>=1.0.0
